Use platform specific clean functions
